Add experimental-evaluator flag
A flag has been added to any package which uses the evaluator, allowing
the old evaluator to be replaced with the new evaluator. This should
allow side-by-side comparisons and testing without changing the
default from the existing evalautor (as the new evalautor will not
be ready until primitive operations are also defined).